urldata: switch to uint* types more widely
In particular, it turns 'unsigned long' into 'uint32_t' since the code
needs to build and run just as fine on Windows which has 32 bit longs,
so we know the code works with 32 bit field versions.

This makes Curl_easy 56 bytes smaller on my 64 bit Linux (maximized
build).

+++ lib/setopt.c
@@ -269,9 +269,9 @@ static CURLcode httpauth(struct Curl_easy *data, bool proxy,
       return CURLE_NOT_BUILT_IN; /* no supported types left! */
   }
   if(proxy)
-    data->set.proxyauth = auth;
+    data->set.proxyauth = (uint32_t)auth;
   else
-    data->set.httpauth = auth;
+    data->set.httpauth = (uint32_t)auth;
   return CURLE_OK;
 }
 #endif /* !CURL_DISABLE_HTTP || !CURL_DISABLE_PROXY */
